requirements.txt

weaviate-client>=3.26.7,<4.0.0
langchain>=0.2.0
langchain_openai
sse_starlette
langserve>=0.0.42
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pandas>=2.2.2
